+++ packages/feedback/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.4](https://github.com/Availity/availity-react/compare/@availity/feedback@5.0.3...@availity/feedback@5.0.4) (2019-08-15)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [5.0.3](https://github.com/Availity/availity-react/compare/@availity/feedback@5.0.2...@availity/feedback@5.0.3) (2019-08-15)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [5.0.2](https://github.com/Availity/availity-react/compare/@availity/feedback@5.0.1...@availity/feedback@5.0.2) (2019-08-14)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [5.0.1](https://github.com/Availity/availity-react/compare/@availity/feedback@5.0.0...@availity/feedback@5.0.1) (2019-08-13)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
 # [5.0.0](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.21...@availity/feedback@5.0.0) (2019-08-13)